Removed original WCAG text
I'm still assuming (hoping) the original WCAG text will get incorporated through an include so we won't have to include it. It's in the issue because having all of the text there for group understanding of the current requirement with quick comparison of the non-web interpretation is helpful.  

I also removed the bullet markup (not needed) and extra blank line - simple editorials. The rest is perfect.