Update introduction.md - remove repeated therefore in penultimate sentence
Minor editorial in Guidance section, penultimate paragraph.  Therefore used to start the last 2 sentences. 

Suggest a minor reword (substitute "as a result" for the first therefore). 

While WCAG 2 was designed to be technology-neutral, it assumes the presence of a “user agent” such as a browser, media player, or assistive technology as a means to access web content. As a result, the application of WCAG 2 to documents and software in non-web contexts necessitates some interpretation in order to determine how the intent of each WCAG 2 success criterion could be met in these different contexts of use. Therefore, the bulk of the Task Force's work involved evaluating how each WCAG 2 success criterion would apply in the context of non-web ICT, if it were applied to non-web ICT.