Remove the sole link to a term
This PR removes a glossary link for the word 'role'. The reason for this PR is for consistency. There are quite a few other terms throughout the 'problematic' section that have glossary definitions, yet for no apparent reason only this one word had a link.

I believe this change is editorial. However, @maryjom if you feel this would be a substantive change, feel free to reject this PR or have it reviewed.