[Editorial] Add precondition scoping to software on platforms with a keyboard interface service
Per issue #740, not all non-web software runs on a platform with device-independent keyboard interface services. In a Web context this is assumed (since that is viewed in a software stack with a browser layer and an underlying platform that has such services). When applied to all non-web software, this is not always the case that there are such software layers.