2.1.1 Keyboard: Need interpretation for non-web documents
See Issue #752. The WCAG2ICT TF made a resolution in the 7 August meeting to add back in for non-web documents this applies as written. 

This is the same text the published WCAG2ICT note had. The the changes made by the editors neglected to keep the existing text for non-web documents when changes were made on applicability to non-web software.