Update comments-by-guideline-and-success-criterion.md
This is a reimplementation of PR #633 using the current main branch as the original source which was approved by the TF back in May but not incorporated because we hadn't solved non-web documents yet. This new PR is to address issue #627 - for both non-web documents and software since there are problems applying this SC in both cases.