Proposal 3: A variant using Gregg and Mike's input
Gregg, Mike, and I had an email thread to develop potential language. The alternate requirement isn't exactly as Gregg had it but the note is what was proposed. I'm trying to keep this more closely aligned to WCAG2ICT's writing style which doesn't use the precondition style used in the EN 301 549.